requirements.txt
uvicorn
fastapi
pymysql
sqlalchemy
alembic
pyjwt
apscheduler
passlib
bcrypt
jinja2
cachetools
mysqlclient